requirements.txt
Django==3.1.5
djangorestframework==3.12.2
django-filter~=2.4.0
requests~=2.22.0